Reorganize factories related to files
