+++ app/Services/Image/RecipeImageProcessor.php
@@ -2,7 +2,7 @@
 
 namespace App\Services\Image;
 
-use App\Services\File\SaverFactory;
+use App\Services\File\Factories\SaverFactory;
 use App\Services\Image\Factories\ImageFactory;
 use App\Services\Interfaces\Image\ImageProcessorInterface;
 use Illuminate\Http\UploadedFile;
