+++ app/Services/Interfaces/Image/ImageProcessorInterface.php
@@ -2,7 +2,7 @@
 
 namespace App\Services\Interfaces\Image;
 
-use App\Services\File\SaverFactory;
+use App\Services\File\Factories\SaverFactory;
 use App\Services\Image\Factories\ImageFactory;
 use Illuminate\Http\UploadedFile;
 use Intervention\Image\ImageManager;
