+++ tests/Unit/Services/Image/RecipeImageProcessorTest.php
@@ -2,8 +2,8 @@
 
 namespace Tests\Unit\Services\Image;
 
+use App\Services\File\Factories\SaverFactory;
 use App\Services\File\Saver;
-use App\Services\File\SaverFactory;
 use App\Services\Image\Factories\ImageFactory;
 use App\Services\Image\RecipeImageProcessor;
 use App\Services\Interfaces\Image\ImageFactoryInterface;
